+++ NEWS.md
@@ -3,12 +3,13 @@
 front-end to DailyMeasure package
 
 # version 2.4.0
-15th October 2023
+24th October 2023
 
 ## New
 
 * search progress notes in Administration module
-  + requires package striprtf to 'strip' RTF from progress note
+  + requires package `striprtf` to 'strip' RTF from progress note
+  + requires package `readxl` to read Excel spreadsheet of search terms (can also read CSV spreadsheet)
 
 * Supports (and requires) `dMeasure` 0.11.0
   + allows extra encryption key/password for database password